+++ docs/source/News.rst
@@ -4,6 +4,10 @@
 
 This document describes the releases of :mod:`repoze.what`.
 
+After :mod:`repoze.what` 1.0.8 (unreleased)
+===========================================
+
+* Made tests compatible with released :mod:`repoze.who` 1.0.x.
 
 .. _repoze.what-1.0.8: